ISSUE-351 Client: Added form validation to block passwords that match a user's email address.
How I do this is I check the new password against the old and new email addresses that may be inputted in a form. This covers both password changes and when defining passwords at signing up.

The only tricky thing I had to pull off in UserDetailBasicsEditorForm was to add the existing ("init") variable the as the email.old variable on the validation function variables object.

Also took a moment to add some minor flow logic improvements to userDetailFormValidator() so it would be easier to understand.